# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_resume_api_v1_resumes_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_resume_api_v1_resumes_post",
        "type": "object"
      },
...
      "EducationOut": {
        "properties": {
          "degree_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degree Level"
          },
          "degree_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degree Raw"
          },
          "end_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Year"
          },
          "evidence_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence End"
          },
          "evidence_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Start"
          },
          "field_of_study": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Of Study"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "institution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institution"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "start_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Year"
          }
        },
        "required": [
          "id",
          "institution",
          "degree_raw",
          "degree_level",
          "field_of_study",
          "start_year",
          "end_year",
          "evidence_start",
          "evidence_end",
          "position",
          "origin"
        ],
        "title": "EducationOut",
        "type": "object"
      },
...
      "ExperienceOut": {
        "properties": {
          "canonical_title_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Title Id"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "evidence_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence End"
          },
          "evidence_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Start"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_current": {
            "title": "Is Current",
            "type": "boolean"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title_raw": {
            "title": "Title Raw",
            "type": "string"
          }
        },
        "required": [
          "id",
          "company",
          "title_raw",
          "canonical_title_id",
          "level",
          "start_date",
          "end_date",
          "is_current",
          "summary",
          "evidence_start",
          "evidence_end",
          "position",
          "origin"
        ],
        "title": "ExperienceOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatchWeights": {
        "description": "The nominal component weights, surfaced so the ranking is legible.",
        "properties": {
          "experience": {
            "title": "Experience",
            "type": "number"
          },
          "skills": {
            "title": "Skills",
            "type": "number"
          },
          "title": {
            "title": "Title",
            "type": "number"
          }
        },
        "required": [
          "skills",
          "title",
          "experience"
        ],
        "title": "MatchWeights",
        "type": "object"
      },
...
      "ProfileDetail": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationOut"
            },
            "title": "Education",
            "type": "array"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "experiences": {
            "items": {
              "$ref": "#/components/schemas/ExperienceOut"
            },
            "title": "Experiences",
            "type": "array"
          },
          "extracted_at": {
            "format": "date-time",
            "title": "Extracted At",
            "type": "string"
          },
          "extractor_version": {
            "title": "Extractor Version",
            "type": "integer"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "resume": {
            "$ref": "#/components/schemas/ResumeOut"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillOut"
            },
            "title": "Skills",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "resume",
          "full_name",
          "email",
          "phone",
          "extractor_version",
          "extracted_at",
          "reviewed_at",
          "created_at",
          "updated_at",
          "skills",
          "experiences",
          "education"
        ],
        "title": "ProfileDetail",
        "type": "object"
      },
...
      "ProfileMatchesResponse": {
        "properties": {
          "computed_at": {
            "format": "date-time",
            "title": "Computed At",
            "type": "string"
          },
          "filters": {
            "$ref": "#/components/schemas/JobFilters"
          },
          "matched_jobs": {
            "title": "Matched Jobs",
            "type": "integer"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/JobMatchOut"
            },
            "title": "Matches",
            "type": "array"
          },
          "profile": {
            "$ref": "#/components/schemas/MatchProfileSummary"
          },
          "returned": {
            "title": "Returned",
            "type": "integer"
          },
          "total_active_jobs": {
            "title": "Total Active Jobs",
            "type": "integer"
          },
          "weights": {
            "$ref": "#/components/schemas/MatchWeights"
          }
        },
        "required": [
          "profile",
          "filters",
          "weights",
          "total_active_jobs",
          "matched_jobs",
          "returned",
          "matches",
          "computed_at"
        ],
        "title": "ProfileMatchesResponse",
        "type": "object"
      },
      "ProfileUpdate": {
        "description": "Full-document replacement. `reviewed` is the document state, not an\nevent: true sets reviewed_at (now, if not already set), false clears\nit \u2014 editing after review naturally un-reviews by sending false.",
        "properties": {
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationIn"
            },
            "title": "Education",
            "type": "array"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "experiences": {
            "items": {
              "$ref": "#/components/schemas/ExperienceIn"
            },
            "title": "Experiences",
            "type": "array"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "reviewed": {
            "default": false,
            "title": "Reviewed",
            "type": "boolean"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillIn"
            },
            "title": "Skills",
            "type": "array"
          }
        },
        "required": [
          "skills",
          "experiences",
          "education"
        ],
        "title": "ProfileUpdate",
        "type": "object"
      },
...
      "ResumeOut": {
        "properties": {
          "extracted_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Text"
          },
          "file_size": {
            "title": "File Size",
            "type": "integer"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "parser_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Name"
          },
          "parser_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Version"
          },
          "uploaded_at": {
            "format": "date-time",
            "title": "Uploaded At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "file_size",
          "page_count",
          "parser_name",
          "parser_version",
          "uploaded_at",
          "extracted_text"
        ],
        "title": "ResumeOut",
        "type": "object"
      },
...
      "SkillOut": {
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "evidence_snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Snippet"
          },
          "evidence_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Start"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "occurrences": {
            "title": "Occurrences",
            "type": "integer"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "technology_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technology Id"
          }
        },
        "required": [
          "id",
          "technology_id",
          "label",
          "confidence",
          "occurrences",
          "evidence_snippet",
          "evidence_start",
          "origin"
        ],
        "title": "SkillOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Xcelsior API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/profiles/{profile_id}": {
      "delete": {
        "description": "Delete the profile and its uploaded resume (user-owned data).",
        "operationId": "remove_profile_api_v1_profiles__profile_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Profile",
        "tags": [
          "profiles"
        ]
      },
      "get": {
        "operationId": "get_profile_api_v1_profiles__profile_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Profile",
        "tags": [
          "profiles"
        ]
      },
      "put": {
        "description": "Full-document reconcile: id = keep/update, no id = create as\nmanual, absent = delete. One transaction \u2014 the review screen's Save.",
        "operationId": "put_profile_api_v1_profiles__profile_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Profile",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles/{profile_id}/matches": {
      "get": {
        "description": "Rank active jobs against this profile (M9). The profile must belong to\nthe signed-in user (404 otherwise) and be reviewed (409 otherwise \u2014\ninvariant #9: an unreviewed extraction is not usable). Accepts the same\n`JobFilters` the market surface does; the candidate job pool is the\nshared predicate, so the counts agree with search.",
        "operationId": "profile_matches_api_v1_profiles__profile_id__matches_get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Free-text query (websearch syntax)",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text query (websearch syntax)",
              "title": "Q"
            }
          },
          {
            "description": "Canonical title family slug",
            "in": "query",
            "name": "title",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Canonical title family slug",
              "title": "Title"
            }
          },
          {
            "description": "City slug, e.g. san-francisco-ca",
            "in": "query",
            "name": "location",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "City slug, e.g. san-francisco-ca",
              "title": "Location"
            }
          },
          {
            "in": "query",
            "name": "radius_miles",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Radius Miles",
              "type": "integer"
            }
          },
          {
            "description": "Technology slugs (AND)",
            "in": "query",
            "name": "tech",
            "required": false,
            "schema": {
              "default": [],
              "description": "Technology slugs (AND)",
              "items": {
                "type": "string"
              },
              "title": "Tech",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "arrangement",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "onsite",
                    "hybrid",
                    "remote"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Arrangement"
            }
          },
          {
            "in": "query",
            "name": "experience_level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "entry",
                    "mid",
                    "senior",
                    "staff_plus"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Experience Level"
            }
          },
          {
            "description": "Annual USD floor",
            "in": "query",
            "name": "salary_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Annual USD floor",
              "title": "Salary Min"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileMatchesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Profile Matches",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles/{profile_id}/reextract": {
      "post": {
        "description": "Rebuild the profile from the stored resume text, discarding all\nedits \u2014 an explicit user action (ADR-005), useful after taxonomy\ngrowth. The resume bytes and text are untouched.",
        "operationId": "reextract_profile_api_v1_profiles__profile_id__reextract_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reextract Profile",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/resumes": {
      "post": {
        "description": "Upload a PDF resume: parse, extract, and return the structured\nprofile in one synchronous call (the pipeline is deterministic rules,\nwell under a second). The resume and profile are owned by the uploading\nuser. Re-uploading identical bytes is idempotent and returns the user's\nexisting profile untouched \u2014 edits survive.",
        "operationId": "upload_resume_api_v1_resumes_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resume_api_v1_resumes_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Resume",
        "tags": [
          "profiles"
        ]
      }
    },
...
  }
}